Allow setting epgpManager when property is not defined in database.
On the production server all existing guilds do not have the database property epgpManager defined.
In the message hande a test is performed if the epgpManager exists. This result it not being able to set/get the epgpManager value.

The only option is to remove and add the guild again which is not nice.